requirements.txt
lsbeautifulsoup4==4.9.1
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cssselect2==0.3.0
defusedxml==0.6.0
html5lib==1.1
idna==2.10
Pillow==7.2.0
pycparser==2.20
Pyphen==0.9.5
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tinycss2==1.0.2
urllib3==1.25.9
WeasyPrint==51
webdriver-manager==3.2.1
webencodings==0.5.1
wkhtmltopdf==0.2